Add files via upload
The purpose of this project was the implementation of audio effects. Specifically, Flanger, Compression, Reverberation, Tremolo, Overdrive, Panning, Guitar Distortion, Echo, Fuzz, Chorus, Vibrato, Wah-Wah, Convolutional Reverberation and Expansion audio effects were implemented. The script "main.m" constitutes the main script. This script accepts an audio file as input and lets the user choose between serial and parallel processing. If serial processing is chosen, the program asks the user to choose which audio effects to apply on the input file and applies them one after the other on the input audio file. The output file contains all the audio effects applied serially on the input file.  If parallel processing is chosen, all audio effects are applied on the input file in parallel, thus resulting in 14 output files. Each file contains the result of an audio effect applied on the input file.